requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]>=3.1
supabase
python-dotenv
pydantic[email]